requirements.txt
alembic==1.11.1
asyncpg==0.28.0
fastapi==0.100.0
sqlmodel==0.0.8
uvicorn==0.22.0
dotmap==1.3.*
fastapi-limiter==0.1.6
fastapi-redis-cache==0.2.5
python-dotenv==1.0.*
requests==2.31.*
pyjwkest==1.4.*
prometheus-fastapi-instrumentator==6.1.*
opentelemetry-distro==0.43b0
opentelemetry-instrumentation-logging==0.43b0
opentelemetry-instrumentation-urllib==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-instrumentation-urllib3==0.43b0
opentelemetry-exporter-otlp==1.22.*
